+++ Src/PerformanceCollector/Shared/Implementation/PerformanceCollectorEventSource.cs
@@ -88,6 +88,13 @@
             this.WriteEvent(8, misconfiguredCountersCount, overallConfiguredCountersCount, e, this.ApplicationName);
         }
 
+        [Event(15, Level = EventLevel.Verbose,
+            Message = @"Performance counters are unavailable when the application is running under IIS Express. Use EnableIISExpressPerformanceCounters element with a value of 'true' within the Performance Collector Module element to override this behavior.")]
+        public void RunningUnderIisExpress(string applicationName = "dummy")
+        {
+            this.WriteEvent(15, this.ApplicationName);
+        }
+
         #endregion
 
         #region Data reading - success