+++ Test/PerformanceCollector/FunctionalTests/TestApps/TestApp45/Global.asax.cs
@@ -29,7 +29,10 @@ namespace TestApp45
             GlobalConfiguration.Configure(WebApiConfig.Register);
 
             var module = new PerformanceCollectorModule();
-            
+
+            // we're running under IIS Express, so override the default behavior designed to prevent a deadlock
+            module.EnableIISExpressPerformanceCounters = true;
+
             // set test-friendly timings
             var privateObject = new PrivateObject(module);
             privateObject.SetField("collectionPeriod", TimeSpan.FromMilliseconds(10));